requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.2
lightgbm==4.1.0
shap==0.43.0
joblib==1.3.2
numpy==1.24.4
pandas==2.0.3

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# Statistical Analysis
scipy==1.11.4
statsmodels==0.14.0
imbalanced-learn==0.11.0

# API Documentation
pydantic[email]==2.5.0

# Logging and Monitoring
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production Dependencies
gunicorn==21.2.0